extract serialize_points to influxdb_common.rs
