+++ metrics/src/influxdb_v2.rs
@@ -1,14 +1,11 @@
 use {
     crate::{
         datapoint::DataPoint,
+        influxdb_common::serialize_points,
         metrics::{MetricsWriter, HOST_ID},
     },
     log::*,
-    std::{
-        env,
-        fmt::Write,
-        time::{Duration, UNIX_EPOCH},
-    },
+    std::{env, time::Duration},
 };
 
 #[derive(Debug)]
@@ -95,41 +92,6 @@ impl MetricsWriter for Writer {
             warn!("tried to submit points to influxdb v2 but no url was set");
         }
     }
-}
-
-pub fn serialize_points(points: &Vec<DataPoint>, host_id: &str) -> String {
-    const TIMESTAMP_LEN: usize = 20;
-    const HOST_ID_LEN: usize = 8; // "host_id=".len()
-    const EXTRA_LEN: usize = 2; // "=,".len()
-    let mut len = 0;
-    for point in points {
-        for (name, value) in &point.fields {
-            len += name.len() + value.len() + EXTRA_LEN;
-        }
-        for (name, value) in &point.tags {
-            len += name.len() + value.len() + EXTRA_LEN;
-        }
-        len += point.name.len();
-        len += TIMESTAMP_LEN;
-        len += host_id.len() + HOST_ID_LEN;
-    }
-    let mut line = String::with_capacity(len);
-    for point in points {
-        let _ = write!(line, "{},host_id={}", &point.name, host_id);
-        for (name, value) in point.tags.iter() {
-            let _ = write!(line, ",{name}={value}");
-        }
-
-        let mut first = true;
-        for (name, value) in point.fields.iter() {
-            let _ = write!(line, "{}{}={}", if first { ' ' } else { ',' }, name, value);
-            first = false;
-        }
-        let timestamp = point.timestamp.duration_since(UNIX_EPOCH);
-        let nanos = timestamp.unwrap().as_nanos();
-        let _ = writeln!(line, " {nanos}");
-    }
-    line
 }
 
 pub fn get_config() -> Config {
